# Conflicts:
#	client/package-lock.json
#	client/package.json
#	client/src/App.jsx

+++ server/src/main/resources/db/migration/V2__create_forecast_table.sql
@@ -2,7 +2,7 @@ CREATE TABLE forecasts
 (
     id               UUID    NOT NULL,
     creation_time    TIMESTAMP WITHOUT TIME ZONE,
-    forecasted_usage INTEGER NOT NULL,
+    forecasted_usage DECIMAL NOT NULL,
     forecast_date    TIMESTAMP WITHOUT TIME ZONE,
     CONSTRAINT pk_forecasts PRIMARY KEY (id)
 );
